+++ init_web_v2.go
@@ -156,8 +156,8 @@ func initWebV2Actions(app *App) {
 	m.Get("/v3/create_atomic_swap_requests/{id}", handlers.GetCreateAtomicSwapRequests)
 	m.Get("/v3/create_poll_requests", handlers.GetCreatePollRequests)
 	m.Get("/v3/create_poll_requests/{id}", handlers.GetCreatePollRequests)
-	//m.Get("/v3/kyc_recovery_requests", handlers.GetKYCRecoveryRequests)
-	//m.Get("/v3/kyc_recovery_requests/{id}", handlers.GetKYCRecoveryRequests)
+	m.Get("/v3/kyc_recovery_requests", handlers.GetKYCRecoveryRequests)
+	m.Get("/v3/kyc_recovery_requests/{id}", handlers.GetKYCRecoveryRequests)
 
 	m.Get("/v3/key_values", handlers.GetKeyValueList)
 	m.Get("/v3/key_values/{key}", handlers.GetKeyValue)