+++ web_v2/handlers/get_movements.go
@@ -32,7 +32,7 @@ func GetMovements(w http.ResponseWriter, r *http.Request) {
 }
 
 // GetMovements returns the list of participant effects with related resources
-func (h *getHistory) GetMovements(request *requests.GetHistory) (regources.ParticipantsEffectsResponse, error) {
+func (h *getHistory) GetMovements(request *requests.GetHistory) (regources.ParticipantsEffectListResponse, error) {
 	q := h.ApplyFilters(request, h.EffectsQ).Movements()
 	result, err := h.SelectAndPopulate(request, q)
 	if err != nil {