+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Fixed
 
 * Sale participation response rendering
-
+* Names of multiple resources responses changed (e.g. `AccountsResponse` -> `AccountListResponse`)
 
 ## 3.3.1-x.0