+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+### Added
+
+* `InitiateKYCRecovery`, `CreateKYCRecoveryRequest`  operations ingestion
+* `/v3/kyc_recovery_requests` endpoint to get filtered list of KYC Recovery requests
+* `/v3/kyc_recovery_requests/{id` endpoint to get KYC Recovery request by id
+
 ### Fixed
 
 * Sale participation response rendering