+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * `/v3/sales/{id}/relationships/whitelist` endpoint to get list of accounts explicitly granted 
 permission to participate in sale.
 * `/v3/sales/{id}/relationships/participation` endpoint to get list of sale participations
+* `/v3/account/{id}/converted_balances/{asset_code}` endpoint to get collection of converted balances
 * Sale participation on sale close ingestion
 * Account specific rules ingestion
 * Sale version ingestion