+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Removed
 
 * handling `/atomic_swap_bids` request
+* Names of multiple resources responses changed (e.g. `AccountsResponse` -> `AccountListResponse`)
 
 ## 3.3.1-x.0
 
@@ -26,6 +27,7 @@ permission to participate in sale.
 * Account specific rules ingestion
 * Sale version ingestion
 * Bumped up ingest version
+* `limits_with_stats` to `account` resource
 
 ## 3.3.0